Gracefully handle deserialization of Joda types
that were serialized without first registering
the JodaModule. Under such a situation, objects
containing Joda types would be serialized with
Jackson's default serialization behavior which
is to use all the object's public getters to
inform serialization. This results in rather
verbose output, but no errors or warnings logged.
More critically, the resulting json cannot be
deserialized even after registering the JodaModule.
Neglecting to register the JodaModule is an easy trap
to fall into because doing so was not required in
Jackson 1.x, which was a monolith that supported
Joda types out of the box. Jackson 2.x was broken
out into modules. Anyone serializing objects with
Joda types who upgraded to Jackson 2.x without
realizing this might have inadvertently serialized
and persisted objects which are undeserializable.
This fix adds support for deserializing all
Joda types which have been serialized with
Jackson's default all-public-getters serializer.